remove unnecessary `alloc` feature
