+++ bee-common/bee-common/Cargo.toml
@@ -11,7 +11,6 @@ keywords = ["iota", "tangle", "bee", "framework", "common"]
 homepage = "https://www.iota.org"
 
 [features]
-alloc = []
 io = []
 
 [build-dependencies]
@@ -21,4 +20,4 @@ autocfg = "1.0"
 bee-common-derive = { path = "../bee-common-derive" }
 
 [dev-dependencies]
-bee-common = { path = ".", features = ["alloc", "io"] }
+bee-common = { path = ".", features = ["io"] }
